handlers/reply.py
from telegram import Update
from telegram.ext import ContextTypes
from config.log import setup_logger 

logger = setup_logger()

async def reply(update: Update, context: ContextTypes):
    logger.info("reply here!")