requirements.txt
amocrm-api==2.6.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.5
billiard==4.2.2
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-pkg==0.14
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
deep-translator==1.11.4
Django==5.2.7
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-parler==2.3
django-parler-rest==2.2
django-rosetta==0.10.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
drf-yasg==1.21.10
firebase-admin==6.8.0
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
msgpack==1.1.0
ngrok==1.4.0
packaging==25.0
parler==1.0.1
pi==0.1.2
pillow==11.2.1
polib==1.2.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.11
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
sentry-sdk==2.35.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==7.2
